greatly simplify compat plugins, error on v1, support als
