+++ alchemy/src/cloudflare/bundle/build-failures.ts
@@ -53,17 +53,3 @@ export function isBuildFailure(err: unknown): err is esbuild.BuildFailure {
     "warnings" in err
   );
 }
-
-/**
- * Returns true if the error has a cause that is like an esbuild BuildFailure object.
- */
-export function isBuildFailureFromCause(
-  err: unknown,
-): err is { cause: esbuild.BuildFailure } {
-  return (
-    typeof err === "object" &&
-    err !== null &&
-    "cause" in err &&
-    isBuildFailure(err.cause)
-  );
-}
