+++ alchemy/test/cloudflare/bundle-handler-als.ts
@@ -1,0 +1,7 @@
+import hooks from "node:async_hooks";
+
+export default {
+  async fetch(): Promise<Response> {
+    return new Response(typeof hooks.AsyncLocalStorage);
+  },
+};
